+++ th2_data_services/provider/v5/commands/http.py
@@ -393,20 +393,22 @@ class GetMessageById(IHTTPProvider5Command, ProviderAdaptableCommand):
         MessageNotFound: If message by id wasn't found.
     """
 
-    def __init__(self, id: str, use_stub: bool = False):
+    def __init__(self, id: str, use_stub: bool = False, only_raw: bool = False):
         """GetMessageById constructor.
 
         Args:
             id: Message id.
             use_stub: If True the command returns stub instead of exception.
+            only_raw: If True the command returns message without parsed body.
         """
         super().__init__()
         self._id = id
         self._stub_status = use_stub
+        self._only_raw = only_raw
 
     def handle(self, data_source: HTTPProvider5DataSource) -> dict:  # noqa: D102
         api: HTTPProvider5API = data_source.source_api
-        url = api.get_url_find_message_by_id(self._id)
+        url = api.get_url_find_message_by_id(self._id, self._only_raw)
 
         # LOG         logger.info(url)
 
@@ -437,22 +439,25 @@ class GetMessagesById(IHTTPProvider5Command, ProviderAdaptableCommand):
         MessageNotFound: If any message by id wasn't found.
     """
 
-    def __init__(self, ids: List[str], use_stub: bool = False):
+    def __init__(self, ids: List[str], use_stub: bool = False, only_raw: bool = False):
         """GetMessagesById constructor.
 
         Args:
             ids: Message id list.
             use_stub: If True the command returns stub instead of exception.
+            only_raw: If True the command returns messages without parsed body.
         """
         super().__init__()
         self._ids: ids = ids
         self._stub_status = use_stub
-
+        self._only_raw = only_raw
+        
     def handle(self, data_source: HTTPProvider5DataSource) -> List[dict]:  # noqa: D102
         result = []
         for message_id in self._ids:
             message = GetMessageById(
                 message_id,
+                only_raw=self._only_raw,
                 use_stub=self._stub_status,
             ).handle(data_source)
             if message is None: