+++ release_notes.md
@@ -379,7 +379,8 @@ This release implements performance bug fixes and provides Data object cache fil
 ## Feature
 1. [TH2-4601] Added `clear_cache` method to data object to remove cache file.
 2. [TH2-4787] Added DatetimeConverter and ProtobufTimestampConverter classes from v2.
-3. [TH2-4751] Add event/message field resolvers.
+3. [TH2-4849] Added only_raw argument to v5 GetMessageById command.
+4. [TH2-4751] Add event/message field resolvers.
    - Global variables added
      - EVENT_FIELDS_RESOLVER: EventFieldsResolver
      - MESSAGE_FIELDS_RESOLVER: MessageFieldsResolver